re-enable minBE128/maxBE128 optimizations
