+++ page_min_amd64.go
@@ -20,5 +20,5 @@ func minFloat32(data []float32) float32
 //go:noescape
 func minFloat64(data []float64) float64
 
-// go:noescape
-// func minBE128(data []byte) []byte
+//go:noescape
+func minBE128(data [][16]byte) []byte
